+++ cmd/new.go
@@ -459,7 +459,9 @@ func installDependencies() error {
 	var c *exec.Cmd
 	if strings.EqualFold(proj.Runtime.Name(), "nodejs") {
 		command = "npm install"
-		c = exec.Command("npm", "install")
+		// We pass `--loglevel=error` to prevent `npm` from printing warnings about missing
+		// `description`, `repository`, and `license` fields in the package.json file.
+		c = exec.Command("npm", "install", "--loglevel=error")
 	} else {
 		return nil
 	}