+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Improvements
 
+- Minor contributing document improvement.
 - Warnings from `npm` about missing description, repository, and license fields in package.json are
   now suppressed when `npm install` is run from `pulumi new` (via `npm install --loglevel=error`).
 - Depend on newer version of gRPC package in the NodeJS SDK. This version has